* commit 'v0.8.2-69-g97ae828': (638 commits)
  Fix MSVC error C2036: 'void *' : unknown size
  Update version to 0.8.1
  Remove colon as multi-line comment symbol
  Fix vectorized_ops for bento
  BUG: not all histogram memory was aligned correctly
  Add release instruction to update dev docs
  Add mailing list adresses to release instructions
  Use np.intp type for indices
  Fix `ValueError: Buffer dtype mismatch` on win-amd64
  Fix `ValueError: Buffer dtype mismatch` on win-amd64
  Fix `ValueError: Buffer dtype mismatch` on win-amd64
  Fix `ValueError: Buffer dtype mismatch` on win-amd64
  Increase version number to 0.9
  Fix release instruction for making docs
  Update docversion for 0.8
  Update version number for 0.8
  Add release notes for version 0.8
  DOC: Fix typo in remove_small_objects docstring.
  Improve performance of normalization
  Change color of circles for better visibility
  ...

+++ skimage/feature/tests/test_hog.py
@@ -102,7 +102,7 @@ def test_hog_orientations_circle():
     width = height = 100
 
     image = np.zeros((height, width))
-    rr, cc = draw.circle(height/2, width/2, width/3)
+    rr, cc = draw.circle(int(height / 2), int(width / 2), int(width / 3))
     image[rr, cc] = 100
     image = ndimage.gaussian_filter(image, 2)
